add CrossTabChart react component
