+++ apps/test/unit/storage/dataBrowser/DataVisualizerTest.js
@@ -126,75 +126,6 @@ describe('DataVisualizer', () => {
     });
   });
 
-  describe('createPivotTable', () => {
-    let wrapper;
-    beforeEach(() => {
-      wrapper = shallow(<DataVisualizer {...DEFAULT_PROPS} />);
-    });
-
-    it('populates zeroes for all row/column combinations', () => {
-      const records = [
-        {abc: 'a', value: 1},
-        {abc: 'b', value: 2},
-        {abc: 'c', value: 3}
-      ];
-      const expectedPivotData = {
-        chartData: [
-          {abc: 'a', 1: 1, 2: 0, 3: 0},
-          {abc: 'b', 1: 0, 2: 1, 3: 0},
-          {abc: 'c', 1: 0, 2: 0, 3: 1}
-        ],
-        columns: ['abc', 1, 2, 3]
-      };
-      expect(
-        wrapper.instance().createPivotTable(records, 'abc', 'value')
-      ).to.deep.equal(expectedPivotData);
-    });
-
-    it('sorts string columns alphabetically, but with the row column first', () => {
-      const records = [
-        {abc: 'a', value: 'z'},
-        {abc: 'b', value: 'z'},
-        {abc: 'a', value: 'z'},
-        {abc: 'c', value: 'z'},
-        {abc: 'c', value: 'a'}
-      ];
-      const expectedPivotData = {
-        chartData: [
-          {abc: 'a', a: 0, z: 2},
-          {abc: 'b', a: 0, z: 1},
-          {abc: 'c', a: 1, z: 1}
-        ],
-        columns: ['abc', 'a', 'z']
-      };
-      expect(
-        wrapper.instance().createPivotTable(records, 'abc', 'value')
-      ).to.deep.equal(expectedPivotData);
-    });
-
-    it('sorts numeric columns numerically, with the row column first', () => {
-      const records = [
-        {abc: 'a', value: 134},
-        {abc: 'b', value: 134},
-        {abc: 'a', value: 134},
-        {abc: 'c', value: 134},
-        {abc: 'c', value: 18}
-      ];
-      const expectedPivotData = {
-        chartData: [
-          {abc: 'a', 18: 0, 134: 2},
-          {abc: 'b', 18: 0, 134: 1},
-          {abc: 'c', 18: 1, 134: 1}
-        ],
-        columns: ['abc', 18, 134]
-      };
-      wrapper.instance().setState({numericColumns: ['value']});
-      expect(
-        wrapper.instance().createPivotTable(records, 'abc', 'value')
-      ).to.deep.equal(expectedPivotData);
-    });
-  });
-
   describe('parseRecords', () => {
     let wrapper;
     beforeEach(() => {
